requirements.txt
openai>=1.40.0
pandas>=2.2.0
reportlab>=4.2.0
streamlit>=1.36.0
python-dotenv>=1.0.1
